Add recipe directory and baked-beans.html
Created new directory in odin-recipes called "recipes".
Added the baked-beans.html file to recipes directory.
Created HTML boilerplate for baked-beans.html and added both a title
and a h1 element to let users know which page they are accessing
from their browser bar and to give information as to which recipe the
site will contain.