Add HTML boilerplate, title, h1, image
Hamburger-buns.html has been created in the recipes directory. A HTML
boilerplate has been coded for this page and the title has been
assigned as "Bert's Hamburger Buns" so as to inform the user which
page they are visiting in the browser bar.

A h1 element has been added, stating the name of the recipe and an
image of the final product has been added underneath said h1 element.
Both the src and alt attributes have been assigned values so that
the page is accessible.