Add em, h2, p and ul elements
An em element has been added in line 17 to place emphasis on the word
"do" to make the text more accessible for those who use screen
readers.

A h2 element "Ingredients" has been added followed by a p and ul
element. This allows the user to see which ingredients they will be
using to recreate this recipe.