Add h2 and ol elements
A h2 element "Steps" has been added, followed by a ol element
containing the steps that need to be followed in the specified order
to recreate the recipe. The li elements nested in the ol parent
element have been separated using br elements to make the steps
easier to read.